refactor(models): extract displacement-normal shader to the shared lib
The averaged-by-position displacement-direction logic and its two GLSL
chunks were hand-copied between the frontend
(shared/three/sharedDisplacementNormal.ts) and the worker applyTextures
page.evaluate — a prime drift candidate (duplicated shader source).

Move it to asset-processor/lib/displacementNormal.js (THREE injected into
addSharedDisplacementNormal; applyDispNormalDisplacement only touches
material hooks; GLSL strings are the single source; window.modelibrDispNormal
side-effect). The frontend file becomes a thin wrapper that injects the
bundler's THREE so all call sites and the existing Jest test are unchanged;
the render template loads the module and applyTextures calls it via the
window global. Documents it in lib/README.md.
